// frontend/src/lib/prisma-types.ts
export interface Profile {
  id: string;
  uid: string;
  type: string | null;
  username: string | null;
  displayName: string | null;
  firstName: string | null;
  lastName: string | null;
  email: string | null;
  phone: string | null;
  tagline: string | null;
  bio: string | null;
  website: string | null;
  size: string | null;
  category: string | null;
  subcategory: string | null;
  speciality: string | null;
  skills: string[];
  coverage: AppJson.Coverage | null;
  portfolio: AppJson.Portfolio | null;
  visibility: AppJson.VisibilitySettings | null;
  socials: AppJson.SocialMedia | null;
  viber: string | null;
  whatsapp: string | null;
  rate: number | null;
  commencement: string | null;
  experience: number | null;
  contactMethods: string[];
  paymentMethods: string[];
  settlementMethods: string[];
  budget: string | null;
  terms: string | null;
  billing: AppJson.BillingInfo | null;
  image: string | null;
  authorBio: string | null;
  verified: boolean;
  featured: boolean;
  rating: number;
  reviewCount: number;
  stars: AppJson.StarBreakdown | null;
  top: boolean;
  published: boolean;
  isActive: boolean;
  createdAt: string;
  updatedAt: string;
}